refactor : jwt 관련 refactoring
