+++ elcoco/elcoco/src/main/java/com/sku/elcoco/controller/member/MemberApiController.java
@@ -1,5 +1,9 @@
-package com.sku.elcoco.config.jwt;
+package com.sku.elcoco.controller.member;
 
+import com.sku.elcoco.domain.member.dto.MemberLoginRequestDto;
+import com.sku.elcoco.domain.member.dto.MemberRequestDto;
+import com.sku.elcoco.service.member.MemberService;
+import com.sku.elcoco.config.jwt.TokenInfo;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.PostMapping;
